feat: added all missing parsers

+++ internal/parsers/extended_fields/inode_extended_field_reader.go
@@ -32,7 +32,7 @@ func (r *InodeExtendedFieldReader) PreviousFileSize() (uint64, bool) {
 
 func (r *InodeExtendedFieldReader) FinderInfo() ([]byte, bool) {
 	for _, f := range r.fields {
-		if f.Type() == int(types.JObjTypeExtent) {
+		if f.Type() == uint8(types.JObjTypeExtent) {
 			return f.Data(), true
 		}
 	}
